60 FPS. Refactored Turret guns and added more. Projectile size now grows with Scout size.

+++ src/main/java/projectile/ProjectileFactory.java
@@ -89,7 +89,7 @@ public class ProjectileFactory {
 		this.ready = b;
 	}
 
-	public final double getFiringDelay() {
+	public final int getFiringDelay() {
 		return firingDelay;
 	}
